api/pyproject.toml
[tool.poetry]
name = "birdskeras"
version = "0.1.0"
description = ""
authors = ["Luis Villazon Esteban <villazonpersonal@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.121.2"
uvicorn = "^0.38.0"
numpy = "1.26.4"
pillow = "^12.0.0"
#keras = "^3.11.3"
tensorflow = "2.18.0"
#tensorflow-metal = "^1.2.0"
python-multipart = "^0.0.20"
torch = "2.6.0"

keras = "3.8.0"
sqlalchemy = "^2.0.44"
pymysql = "^1.1.2"
pandas = "^2.3.3"
cryptography = "^46.0.3"
mlflow = "^3.6.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
